backend-python/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.30.0
pyjwt[crypto]==2.9.0
httpx==0.28.1
python-dotenv==1.0.1
apscheduler==3.10.4
playwright==1.49.1
websockets==14.1
pydantic==2.10.3
aiosmtplib==5.1.0
